+++ app/Http/Middleware/CheckForMaintenanceMode.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\Http\Middleware;
+
+use Illuminate\Foundation\Http\Middleware\CheckForMaintenanceMode as Middleware;
+
+class CheckForMaintenanceMode extends Middleware
+{
+    protected $except = [
+        //
+    ];
+}